Refactored codebase and added command line interface
The source code is now in the form of the standard structure of a python
package. The rationale is given someone has their pytohn package path
include the code, they can use the modules like any other library.

In addition, there is now a command line interface which allows one to
easily utilize the bace package. See "bace_driver.py" for a demo of how
to use it.

Finally, there is now documentation following the numpy docstrings
format. Human readability was the main reason why this format was chosen
over other options.